refactor(hpsys_rcc): organize clock mux enums into vals::mux submodule
Move clock multiplexer enums (Lpsel, Mpisel, Perisel, Ticksel, Usbsel)
from vals module into vals::mux submodule, following embassy-stm32
naming conventions.

This provides better organization by separating basic clock types
(Sysclk, Pdiv, Dllstg) from clock multiplexer selections.
